JtR: refactor flatpak building
Move all "shared" code to new code units.

+++ john-the-ripper/flatpak/build.sh
@@ -1,28 +1,20 @@
 #!/bin/bash
-if [[ "$TEST" != "yes" ]]; then
-    arch=$(uname -m)
-    text='x'
-    git_tag=$(git describe --dirty=+ --always 2>/dev/null)
 
-    case "$arch" in
-        'x86_64')
-            text='X'
-            ;;
-        'armhf' | 'armv7l')
-            text='a'
-            ;;
-        'aarch64' | 'arm64')
-            text='B'
-            ;;
-        'ppc64le' | 'powerpc64le')
-            text='P'
-            ;;
-        's390x')
-            text='S'
-            ;;
-    esac
-    # View package version
-    echo "s/edge/1.9J1-$git_tag$text/g"
+# Required defines
+TEST=';full;extra;' # Controls how the test will happen
+arch=$(uname -m)
+JTR_BIN='/app/bin/john'
+JTR_CL='/app/bin/john-opencl'
+
+# Show environmen information
+source ../show_info.sh
+
+if [[ -z "$TASK" ]]; then
+    # Set package version
+    git rev-parse --short HEAD 2>/dev/null > My_VERSION.TXT
+
+    # The script that computes the package version
+    source ../package_version.sh
 
     echo ""
     echo "---------------------------- BUILDING -----------------------------"
@@ -42,60 +34,20 @@ if [[ "$TEST" != "yes" ]]; then
         ./configure --disable-native-tests --disable-opencl --with-systemwide --disable-openmp CPPFLAGS="-D_BOXED" && make -s clean && make -sj2 && mv ../run/john ../run/john-non-omp
         ./configure --disable-native-tests --disable-opencl --with-systemwide                  CPPFLAGS="-D_BOXED -DOMP_FALLBACK -DOMP_FALLBACK_BINARY=\"\\\"john-non-omp\\\"\"" && make -s clean && make -sj2
     fi
-    # Remove unused stuff
-    rm -rf ../run/kerberom
-    rm -rf ../run/ztex
 
-    # Workaround for non X86
-    ln -s ../run/john ../run/john-opencl
+    # Workaround for non X86 (non-OpenCL)
+    if [[ ! -f ../run/john-opencl ]]; then
+        ln -s ../run/john ../run/john-opencl
+    fi
 
-else
-    echo ""
-    echo "---------------------------- TESTING -----------------------------"
-    JTR_BIN='/app/bin/john'
-    "$JTR_BIN" --list=build-info
+elif [[ "$TASK" == "test" ]]; then
+    # "---------------------------- TESTING -----------------------------"
 
-    echo '$NT$066ddfd4ef0e9cd7c256fe77191ef43c' > tests.in
-    echo '$NT$8846f7eaee8fb117ad06bdd830b7586c' >> tests.in
-    echo 'df64225ca3472d32342dd1a33e4d7019f01c513ed7ebe85c6af102f6473702d2' >> tests.in
-    echo '73e6bc8a66b5cead5e333766963b5744c806d1509e9ab3a31b057a418de5c86f' >> tests.in
-    echo '$6$saltstring$fgNTR89zXnDUV97U5dkWayBBRaB0WIBnu6s4T7T8Tz1SbUyewwiHjho25yWVkph2p18CmUkqXh4aIyjPnxdgl0' >> tests.in
+    # Adjust the testing environment, import and run some testing
+    source ../disable_formats.sh
 
-    echo
-    echo "====> regex T1 A: 9 lines"
-    "$JTR_BIN" --stdout --regex='[0-2]password[A-C]'
-    echo "====> regex T1 B: 2 lines, 1 special character"
-    "$JTR_BIN" --stdout --regex=ab[öc]
-    echo "====> regex T1 C: 7 lines, 7 special characters, quotation marks"
-    "$JTR_BIN" --stdout --regex="ab[£öçüàñẽ]"
-    echo "====> regex T1 D: 5 lines, 4 special characters, quotation marks"
-    "$JTR_BIN" --stdout --regex='ab(ö|¿|e|¡|!)'
-    echo "====> regex T1 E: 2 lines, 1 special character, vertical bar"
-    "$JTR_BIN" --stdout --regex='ab(ö|c)'
-    echo "====> regex T1 F: 3 lines, 5 special characters, vertical bar"
-    "$JTR_BIN" --stdout --regex='ab(ö,¿|\?,e|¡,!)'
-    echo "====> regex T2: 2 lines, at the end"
-    echo magnum | "$JTR_BIN" -stdout -stdin -regex='\0[01]'
-    echo "====> regex T3 A: 2 lines, at the end, encoding"
-    echo müller | iconv -f UTF-8 -t cp850 | "$JTR_BIN" -inp=cp850 -stdout -stdin -regex='\0[01]'
-    echo "====> regex T3 B: 2 lines, encoding"
-    "$JTR_BIN" -stdout --regex='ab(ö|c)' -target-enc=cp437
-    echo
-    echo "====> T4:"
-    "$JTR_BIN" -test-full=0 --format=nt
-    echo "====> T5:"
-    "$JTR_BIN" -test-full=0 --format=sha256crypt
-    echo "------------------------------------------------------------------"
-    "$JTR_BIN" -test=0 --format=raw*
-    echo "------------------------------------------------------------------"
-
-    echo "====> T10:"
-    "$JTR_BIN" tests.in --format=nt --fork=2
-    echo "====> T11:"
-    "$JTR_BIN" tests.in --format=raw-sha256 --fork=2
-    echo "====> T12-a:"
-    "$JTR_BIN" tests.in --format=sha512crypt --mask=jo?l[n-q]
+    source ../run_tests.sh
 fi
-# Remove the left-over from testing
-rm -f ../run/john.log
-rm -f ../run/john.pot
+source ../clean_package.sh
+
+echo '=================================='
